feat(quay-action): add support for the new backend system
