+++ plugins/quay-actions/README.md
@@ -14,7 +14,7 @@ Run the following command to install the action package in your Backstage projec
 yarn workspace backend add @janus-idp/backstage-scaffolder-backend-module-quay
 ```
 
-## Configuration
+### Installing the action on the legacy backend
 
 1. [Register](https://backstage.io/docs/features/software-templates/writing-custom-actions#registering-custom-actions) the Quay actions by modifying the `packages/backend/src/plugins/scaffolder.ts` file from your project with the following changes:
 
@@ -68,21 +68,38 @@ yarn workspace backend add @janus-idp/backstage-scaffolder-backend-module-quay
    yarn workspace backend add @backstage/integration
    ```
 
-3. Add the Quay actions to your templates, see the [example](./examples/templates/01-quay-template.yaml) file in this repository for complete usage examples
+### Installing the action on the new backend
 
-   ```yaml
-   action: quay:create-repository
-   id: create-quay-repo
-   name: Create quay repo
-   input:
-     baseUrl: https://quay.io
-     token: UW1dLVdCTj8uZWNuIW97K1k0XiBkSmppVU9MYzFT
-     name: foo
-     visibility: public
-     description: This is a foo repository
-     namespace: bar
-     repoKind: image
-   ```
+Add the following to your `packages/backend/src/index.ts` file:
+
+```ts title="packages/backend/src/index.ts"
+const backend = createBackend();
+
+// Add the following line
+backend.add(
+  import('@janus-idp/backstage-scaffolder-backend-module-quay/alpha'),
+);
+
+backend.start();
+```
+
+## Configuration
+
+Add the Quay actions to your templates, see the [example](./examples/templates/01-quay-template.yaml) file in this repository for complete usage examples
+
+```yaml
+action: quay:create-repository
+id: create-quay-repo
+name: Create quay repo
+input:
+  baseUrl: https://quay.io
+  token: UW1dLVdCTj8uZWNuIW97K1k0XiBkSmppVU9MYzFT
+  name: foo
+  visibility: public
+  description: This is a foo repository
+  namespace: bar
+  repoKind: image
+```
 
 ## Usage
