+++ plugins/quay-actions/src/alpha.ts
@@ -1,0 +1,1 @@
+export { scaffolderModuleQuayAction as default } from './module';
