+++ plugins/quay-actions/src/dynamic/alpha.ts
@@ -1,23 +1,8 @@
-import { createBackendModule } from '@backstage/backend-plugin-api';
 import { BackendDynamicPluginInstaller } from '@backstage/backend-plugin-manager';
-import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
 
-import { createQuayRepositoryAction } from '../actions';
+import { scaffolderModuleQuayAction } from '../module';
 
 export const dynamicPluginInstaller: BackendDynamicPluginInstaller = {
   kind: 'new',
-  install: createBackendModule({
-    moduleId: 'scaffolder-backend-quay',
-    pluginId: 'scaffolder',
-    register(env) {
-      env.registerInit({
-        deps: {
-          scaffolder: scaffolderActionsExtensionPoint,
-        },
-        async init({ scaffolder }) {
-          scaffolder.addActions(createQuayRepositoryAction());
-        },
-      });
-    },
-  }),
+  install: scaffolderModuleQuayAction,
 };
